Move file locations and change paths accordingly
